Moved file to respective folder for issue manan-shxrma/dsalgo#27
